seg/seg-requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
apex @ file:///mnt/bn/lianghuidata/apex
asttokens==2.4.1
attrs==23.1.0
Automat==22.10.0
bidict==0.23.1
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
buildtools==1.0.6
causal-conv1d==1.1.0
certifi @ file:///croot/certifi_1720453481653/work/certifi
cffi==1.16.0
chardet==5.2.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.8.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.1
furl==2.1.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.65.0
h11==0.14.0
huggingface-hub==0.23.4
hyperlink==21.0.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
importlib_metadata==8.0.0
importlib_resources==6.4.0
incremental==22.10.0
iniconfig==2.0.0
ipaddress==1.0.23
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
kiwisolver==1.4.5
-e git+https://github.com/hustvl/Vim@cbbef2c75220161cb63a21239864a310239f8aa0#egg=mamba_ssm&subdirectory=mamba-1p1p1
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mmcls==0.25.0
mmcv-full==1.7.1
mmsegmentation==0.29.1
model-index==0.1.11
msgpack==1.0.8
ninja==1.11.1.1
numpy==1.23.4
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1701735466804/work
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.0
ordered-set==4.1.0
orderedmultidict==1.0.1
oss2==2.17.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathlib2==2.3.7.post1
pathtools==0.1.2
pexpect==4.9.0
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1621287979692/work
platformdirs==4.2.2
pluggy==1.5.0
ply==3.11
prettytable==3.10.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-spy==0.3.14
pyarrow==16.1.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytest==6.2.5
python-consul==1.1.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-etcd==0.4.5
python-socketio==5.11.3
pytz==2023.4
PyYAML==6.0.1
redo==2.0.4
regex==2024.5.15
requests==2.28.2
rich==13.4.2
safetensors==0.4.3
schedule==1.2.2
scipy==1.13.1
sentry-sdk==2.9.0
setproctitle==1.3.3
shortuuid==1.0.13
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
stack-data==0.6.3
thriftpy2==0.5.2
timm==1.0.7
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tox==3.28.0
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
triton==2.1.0
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
virtualenv==20.26.3
watchdog==4.0.1
wcwidth==0.2.13
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
yapf==0.40.1
zipp==3.19.2
zope.interface==6.4.post2
